send-events/requirements.txt
azure-eventgrid
azure-core
azure-identity
azure-servicebus